backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet>=3.0.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP client (for LLM API calls)
httpx==0.27.2

# Data processing
numpy>=1.26.0
pandas>=2.2.0
jieba==0.42.1
scikit-learn>=1.5.0
matplotlib>=3.9.0

# Text embedding (optional - install manually if needed)
# sentence-transformers>=3.0.0

# Image processing
Pillow>=10.4.0

# Audio processing (optional - install manually if needed)
# openai-whisper==20231117

# Environment
python-dotenv==1.0.1

# CORS
pydantic[email]==2.9.2
pydantic-settings==2.5.2
